Fix for text running off screen
If the line was multiple lines, then only the second line was
determining the font size. So in the case where the first line was
longer, it would extend past the image.

Changed the max font size to be different for multiline text, this
should prevent text from taking up too much vertical space when the
phrase was just long enough to be split.