add loading gif while meme loads
There is no reason this shouldn't work, but I when I tried to test this, It turned out (pls no hate) that my internet is too fast. I wasn't able to see what happened when memes were loading because it happened too fast. But, i've used this method before, and it has worked.