Swap pymdownx magic links for bare URLs
Unfortunately portray has a version conflict with pymdown-extensions, and even when I overrode it locally, there were some elements that did not display as desired. Since the usage was relatively minor and limited to the README.md file alone, it was just as easy to replace them with bare URLs.